Files related to embeddings moved to another folder
